tel must be int64 to fit a phone number

+++ models/user.go
@@ -3,8 +3,8 @@ package models
 type User struct {
     Id          int64
     NfcSerial   int64
-    FirstName  string
-    LastName   string
-    Tel         int
+    FirstName   string
+    LastName    string
+    Tel         int64
     Devices     []Device
 }
